LeetCodeChallenge#1
This was something I struggled with for a little because I was thinking on how I could use hash maps to reduce run time so I will be doing more research on it and hopefully this gets me closer to what I plan to achieve in the future. But this code is mainly just checking to see if the array has numbers that add up with the targets value.